requirements.txt
Babel==1.3
Fabric==1.8.3
Flask==0.10.1
Flask-Babel==0.9
-e git+git@github.com:ryanolson/flask-bootstrap3.git@1286edf25470a603f4f9ec14fe951d4f6b2fb960#egg=Flask_Bootstrap3-origin/develop
Flask-Login==0.2.10
Flask-Mail==0.9.0
Flask-OAuth==0.12
Flask-Principal==0.4.0
Flask-SQLAlchemy==1.0
Flask-Script==0.6.7
Flask-Security==1.7.1
-e git+git@github.com:ryanolson/flask-social.git@54d6c1da7fc7a8c5e77a7480d0d87005053ccb00#egg=Flask_Social-origin/develop
Flask-WTF==0.9.5
Jinja2==2.7.2
Mako==0.9.1
MarkupSafe==0.19
SQLAlchemy==0.9.4
WTForms==1.0.5
Werkzeug==0.9.4
alembic==0.6.4
amqp==1.4.5
anyjson==0.3.3
arrow==0.4.2
beautifulsoup4==4.3.2
billiard==3.3.0.17
blinker==1.3
celery==3.1.11
ecdsa==0.11
facebook-sdk==0.4.0
factory-boy==2.3.1
github3.py==0.8.2
gunicorn==18.0
httplib2==0.8
itsdangerous==0.24
kombu==3.0.15
oauth2==1.5.211
paramiko==1.12.3
passlib==1.6.2
psycopg2==2.5.2
py==1.4.20
py-bcrypt==0.4
pycrypto==2.6.1
pymongo==2.7
pytest==2.5.2
python-dateutil==1.5
pytz==2014.2
raven==4.1.1
requests==2.2.1
simplejson==3.4.0
six==1.6.1
speaklater==1.3
stravalib==0.2.2
translitcodec==0.3
units==0.06
uritemplate.py==0.3.0
wsgiref==0.1.2
